Extract draggable popup behavior to DraggablePopup
