+++ src/widgets/dialogs/UserInfoPopup.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "widgets/BaseWindow.hpp"
+#include "widgets/DraggablePopup.hpp"
 #include "widgets/helper/ChannelView.hpp"
 
 #include <pajlada/signals/scoped-connection.hpp>
@@ -16,7 +16,7 @@ class Channel;
 using ChannelPtr = std::shared_ptr<Channel>;
 class Label;
 
-class UserInfoPopup final : public BaseWindow
+class UserInfoPopup final : public DraggablePopup
 {
     Q_OBJECT
 
@@ -31,9 +31,6 @@ public:
 protected:
     virtual void themeChangedEvent() override;
     virtual void scaleChangedEvent(float scale) override;
-    void mousePressEvent(QMouseEvent *event) override;
-    void mouseReleaseEvent(QMouseEvent *event) override;
-    void mouseMoveEvent(QMouseEvent *event) override;
 
 private:
     void installEvents();
@@ -54,24 +51,9 @@ private:
     // The channel the messages are rendered from (e.g. #forsen). Can be a special channel, but will try to not be where possible.
     ChannelPtr underlyingChannel_;
 
-    // isMoving_ is set to true if the user is holding the left mouse button down and has moved the mouse a small amount away from the original click point (startPosDrag_)
-    bool isMoving_ = false;
-
-    // startPosDrag_ is the coordinates where the user originally pressed the mouse button down to start dragging
-    QPoint startPosDrag_;
-
-    // requestDragPos_ is the final screen coordinates where the widget should be moved to.
-    // Takes the relative position of where the user originally clicked the widget into account
-    QPoint requestedDragPos_;
-
-    // dragTimer_ is called ~60 times per second once the user has initiated dragging
-    QTimer dragTimer_;
-
     pajlada::Signals::NoArgSignal userStateChanged_;
 
     std::unique_ptr<pajlada::Signals::ScopedConnection> refreshConnection_;
-
-    std::shared_ptr<bool> hack_;
 
     struct {
         Button *avatarButton = nullptr;
